HCK-13303: script generation options - add browser endpoint
