+++ forward_engineering/api/getExternalBrowserUrl.js
@@ -1,0 +1,9 @@
+const reApi = require('../../reverse_engineering/api.js');
+
+async function getExternalBrowserUrl(connectionInfo, logger, cb, app) {
+	reApi.getExternalBrowserUrl(connectionInfo, logger, cb, app);
+}
+
+module.exports = {
+	getExternalBrowserUrl,
+};
